# fix: Speed up group creation by using hash buckets