The achieved speedup by this patch is approx. 3x-120x depending on workload, according to the benchmark script.

By using a `Map`, we avoid array searching entirely in most cases, especially when using simple keys (strings, numbers, booleans etc.). For complex keys, we may still need to fall back to linear-time search when faced with an extreme amount of `JSON.stringify()` collisions.

Thanks to @Nisgrak who came up with the initial idea.

Closes #95