@meyfa 2023-03-29
**Benchmark summary**

The following results were compiled by running `npm run benchmark` on main, and on this PR.

| suite | main ops/s | PR ops/s | speedup |
| ----- | -----------| ----------| ---------|
| remainder-10k | 642.48 | 4731.43 | 736% |
| remainder-100k | 59.75 | 347.92 | 582% |
| object-attribute | 823.09 | 2560.08 | 311% |
| object-callback | 836.57 | 2567.18 | 307% |
| complex-key-100 | 169.97 | 10601.36 | 6,237% |
| complex-key-1000 | 4.65 | 565.48 | 12,160% |

<details>
<summary>Detailed benchmark results</summary>

Run on main:

```
Benchmarking remainder-10k (1000 samples)
total: 1556.470ms, average: 1.556ms, ops/s: 642.48

Benchmarking remainder-100k (100 samples)
total: 1673.745ms, average: 16.737ms, ops/s: 59.75

Benchmarking object-attribute (1000 samples)
total: 1214.929ms, average: 1.215ms, ops/s: 823.09

Benchmarking object-callback (1000 samples)
total: 1195.357ms, average: 1.195ms, ops/s: 836.57

Benchmarking complex-key-100 (200 samples)
total: 1176.709ms, average: 5.884ms, ops/s: 169.97

Benchmarking complex-key-1000 (20 samples)
total: 4303.748ms, average: 215.187ms, ops/s: 4.65
```

Run on PR:

```
Benchmarking remainder-10k (1000 samples)
total: 211.353ms, average: 0.211ms, ops/s: 4731.43

Benchmarking remainder-100k (100 samples)
total: 287.423ms, average: 2.874ms, ops/s: 347.92

Benchmarking object-attribute (1000 samples)
total: 390.613ms, average: 0.391ms, ops/s: 2560.08

Benchmarking object-callback (1000 samples)
total: 389.533ms, average: 0.390ms, ops/s: 2567.18

Benchmarking complex-key-100 (200 samples)
total: 18.866ms, average: 0.094ms, ops/s: 10601.36

Benchmarking complex-key-1000 (20 samples)
total: 35.368ms, average: 1.768ms, ops/s: 565.48
```

</details>